Django/requirements.txt
Django==4.1.1
Pillow==9.2.0
django_filter==22.1
djangorestframework==3.13.1
django-rest-swagger==2.2.0
django_cors_headers==3.11.0
factory-boy==3.2.1

django-filter~=22.1
Faker~=15.1.1